+++ controllers/index.js
@@ -1,0 +1,15 @@
+const {
+  listContacts,
+  getContactById,
+  addContact,
+  updateContact,
+  removeContact,
+} = require("./contacts.js");
+
+module.exports = {
+  listContacts,
+  getContactById,
+  addContact,
+  updateContact,
+  removeContact,
+};
